[Refactor] new module for territory embed
